+++ src/lib/InputInternal.tsx
@@ -27,8 +27,8 @@ const InputInternal = <T extends FieldValues>(props: InputProps<T>) => {
     className,
     style,
     innerRef,
-    minlength,
-    maxlength,
+    minLength,
+    maxLength,
   } = props;
   const { name, id } = useSafeNameId(props.name, props.id);
   const focusHandler = useMarkOnFocusHandler(markAllOnFocus);
@@ -59,8 +59,8 @@ const InputInternal = <T extends FieldValues>(props: InputProps<T>) => {
         }}
         min={rangeMin}
         max={rangeMax}
-        minLength={minlength}
-        maxLength={maxlength}
+        minLength={minLength}
+        maxLength={maxLength}
         rows={textAreaRows}
         multiple={multiple}
         disabled={formDisabled || disabled}