+++ src/lib/types/CommonInputProps.ts
@@ -81,13 +81,13 @@ interface CommonInputProps<T extends FieldValues, TRenderAddon = unknown> {
    * @example <Input<T> type="text" minlength={0} />
    */
 
-  minlength?: number;
+  minLength?: number;
   /**
    * Component prop that represents the maxlength attribute for the input element
    * @type {number}
    * @example <Input<T> type="text" maxlength={10} />
    */
-  maxlength?: number;
+  maxLength?: number;
 }
 
 export { CommonInputProps };